+++ src/exports/apply/sortBy.ts
@@ -1,0 +1,1 @@
+import '../../extensions/sortBy/apply.js';