+++ src/extensions/map/pipeable.ts
@@ -1,0 +1,3 @@
+export const map = <T>(...args: Parameters<Array<T>['map']>) => {
+  return (input: ReadonlyArray<T>) => Array.prototype.map.bind(input)(...args);
+};