+++ src/first/index.ts
@@ -1,1 +1,0 @@
-import './first.js';