+++ src/toRecord/index.ts
@@ -1,1 +1,0 @@
-import './toRecord.js';