+++ src/types/Callback.ts
@@ -1,0 +1,5 @@
+export type Callback<T, U> = (
+  element: T,
+  index: number,
+  collection: ReadonlyArray<T>
+) => U;