+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+# [3.0.0-beta.2](https://github.com/laurence79/ts-array-extensions/compare/3.0.0-beta.1...3.0.0-beta.2) (2025-01-17)
+
+
+### Bug Fixes
+
+* callback should be optional on first() and last() ([c1cc968](https://github.com/laurence79/ts-array-extensions/commit/c1cc9684bb00f9e95d964492e8a2b55a46cf1779))
+
+# [3.0.0-beta.1](https://github.com/laurence79/ts-array-extensions/compare/2.2.4...3.0.0-beta.1) (2025-01-17)
+
+
+### Features
+
+* Add pipes ([#38](https://github.com/laurence79/ts-array-extensions/issues/38)) ([c1bdeff](https://github.com/laurence79/ts-array-extensions/commit/c1bdeff257933eb875b5ed5fb68207d46fa95ab3))
+* Rename addToPrototype to apply ([f4b1c18](https://github.com/laurence79/ts-array-extensions/commit/f4b1c18a8b5b99779eb46f980a1f2eb037798f08))
+
+
+### BREAKING CHANGES
+
+* Async functions removed, new import path to apply to Array prototype
+
+# [3.0.0-alpha.2](https://github.com/laurence79/ts-array-extensions/compare/3.0.0-alpha.1...3.0.0-alpha.2) (2025-01-17)
+
+
+### Features
+
+* Rename addToPrototype to apply ([f4b1c18](https://github.com/laurence79/ts-array-extensions/commit/f4b1c18a8b5b99779eb46f980a1f2eb037798f08))
+
+## [3.0.0-alpha.1](https://github.com/laurence79/ts-array-extensions/compare/2.2.3...3.0.0-alpha.1) (2025-01-16)
+
+
+### Features
+
+* Add pipes ([#38](https://github.com/laurence79/ts-array-extensions/issues/38)) ([c1bdeff](https://github.com/laurence79/ts-array-extensions/commit/c1bdeff257933eb875b5ed5fb68207d46fa95ab3))
+
+
+### BREAKING CHANGES
+
+* Async functions removed, new import path to apply to Array prototype
+
 ## [2.2.4](https://github.com/laurence79/ts-array-extensions/compare/2.2.3...2.2.4) (2025-01-16)
 
 
@@ -7,6 +46,7 @@
 * Individual functions are not exported ([#36](https://github.com/laurence79/ts-array-extensions/issues/36)) ([316690e](https://github.com/laurence79/ts-array-extensions/commit/316690ed8a4dc06a61f923bc47bd4681e26682e0))
 * Missing file extensions on compact and compactMap ([947af23](https://github.com/laurence79/ts-array-extensions/commit/947af237e69aa9054cb876e6e4b40e7eef919f42))
 * Types not included in package ([54f6f44](https://github.com/laurence79/ts-array-extensions/commit/54f6f4416f50d29d4eafbc268349cdff258c4128))
+
 
 ## [2.2.4-next.4](https://github.com/laurence79/ts-array-extensions/compare/2.2.4-next.3...2.2.4-next.4) (2025-01-16)